Add labels as used in the text
The assembly language described in the book might look like:

    LABEL TRAP  x25
          AND   R1,R1,#7
          BRzn  LABEL

To my tastes, that is ambiguous and inconsistent to the reader. I prefer
something like:

    LABEL: TRAP  #x25
           AND   R1,R1,#b111
           BRzn  [LABEL]

We'll try to support both so that code from the text or found online will parse
unmodified while, at the same time, allowing me to be stubborn about trivial
issues of style.