Use figure 7.2 from the second edition
In the text I have, the third edition, the binary code in Fig 7.3 does not match
the source code in figure 7.2. Looking back, the second edition has a different
source in Fig. 7.2 which is a (closer?) match. So, the fixture here is updated
to use the previous edition's source and plan to add an integration test for
code generation using the same fixture.