Add some of an assembler
This got well out of hand. The idea was to avoid having to hand coding interrupt
handlers in machine code, I would (just) build a basic assembler as described in
the text. A week or so later we have (some of) an assembler. As it turns out, it
is a lot of very detailed work, likely more work and more detailed than would be
needed to design, build and debug about five handlers. It has been fun, though!

The assembler is a pretty straightforward two-pass assembler. There are a few
extensions to the language grammar but it should be compatible with the
"specification".

The first pass parses the opcodes and directives in the source and, unlike the
assembler described in the text, holds parsed values in a syntax table for the
second pass. The second could re-parse the source while generating code: it
would save memory and might even use less processor time. But, it seemed more
complicated to make the parser reusable by both the symbol table and code
generation than it to optimize less than 100 KB. The syntax table is backed by a
linear slice, again, for the sake of simplicity. More complicated data
structures would likely just burn clock to save a page.

While testing this code I found that Figure 7.3, the binary machine-code from
listing Fig. 7.3, in the third edition of the text. After quite a bit of
puzzling, I decided that binary output was incorrect in the text. Next, I
searched for the author's errata and found that, in fact, Fig. 7.3 in 3/e looked
similar to the output from 2/e. Don't assume a specification is correct.

There are still quite a few opcodes, a directive to complete and a object file
writer to complete. And, I guess, a loader. This is a lot of work. I'll probably
rewrite all this code a couple more times before ever getting to writing the I/O
routines. At this point, parser6.asm, ie. Fig. 7.2, 2/e, parses and builds.

Squashed commit of: 3a163bb6b2..43c54203b