Wrap operations with source info
Instead of embedding every operation with metadata and having to be careful to
set it when operations are parsed, we wrap all operations a SourceInfo. It
embeds the Operation interface and otherwise delegates to it. Taking a hint from
the error interface, callers can Unwrap to get the original Operation.