Add an object-code loader
The loader takes object code and puts it into memory prior to starting the
machine. There are a few use cases already for the code loader:

  - a new "exec" command for running programs (!)
  - simplification of demo command
  - loading interrupt handlers

Maybe, there ought to be a specialized loader for other things, eg. loading
system monitor code, loaders that verify code(?), more? It does seem like an
interesting abstraction to fall from the design.

The distinction between early/late initialization is more clear in the code and
comments. A simple bool value is passed and the OptionFn can decide what to do.
Still, sequencing and inter-dependencies are going to be worth avoiding.

Squashed commit of a1ce2bfb..f06079d5