Make errors more explicit and easier to unwrap
Error handling is still a dark art. Despite the simple guideline, "errors are
values", working with the kinds and types of errors remains quite a chore.

The use cases here aren't well defined but I am going for a design which:
  - uses errors as values
  - keeps error metadata with the errors
  - presents error messages to users
  - makes debugging info available to programs

This still doesn't quite feel right and the generation side is not yet tidied
towards error types and constants but it does, at least, present to the user
better.

Squashed commit of the following: 7f356d6f..8289384f