Add some developer niceties
Adds:

- a Dockerfile for building a (portable?) binary
- a development guide
- move docs for cleanliness
- a tool for automating development tasks

On the later, some explanation is included in the docs but more, I suspect, is
warranted in history as well. The idea is to avoid adding a dependency on make
or /bin/sh or any other (portable?) tool for command scripting. Although, it is
verbose and a bit weird, it is eminently possible to run commands, parse output,
handle errors, and solve all the problems a scripting language introduces
without adding another necessary tool to the developer's workbench.

For our purposes, it might even be better (simpler, easier to learn, more
stable) to write a bare-essential scripting-language to solve the problem. We
are programmers and when faced with programming problems we ought not be afraid
to solve them with _our own_ programs.

Squashed commit of the following: commit 7701250e..95652d16