Install development dependencies if missing
When using "go run ./internal/tool deps", if stringer or golangci-lint are not
found, try to install them with "go install", for the former, or using the
recommended install script, for the latter. :-(

For golangci-lint, the project specifically recommends the shell script, even
for Windows. Wild! Nevertheless, follow their supported installation process to
avoid problems and because deps is an optional helper command.